third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-dom-height-attr-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'height' attribute of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.getAttribute('height') is "2"
PASS foreignObjectElement.getAttribute('height') is "150"
PASS successfullyParsed is true

TEST COMPLETE
